Refactor custom compostion support
